+++ demo/demo_test.go
@@ -1,9 +1,0 @@
-package demo
-
-import (
-	"testing"
-)
-
-func BenchmarkRun(b *testing.B) {
-	Run(LOOP_COUNT)
-}